+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
   * Added Castable class: [#1634](https://github.com/dartsim/dart/pull/1634)
   * Added spdlog support as underlying logging framework: [#1633](https://github.com/dartsim/dart/pull/1633)
+  * Added MemoryAllocator and CAllocator: [#1636](https://github.com/dartsim/dart/pull/1636)
 
 * Dynamics